devops: remover arquivo de configuracao se existir

+++ scripts/dnit_updater.py
@@ -48,7 +48,8 @@ async def update_deploy(build_name: str, file: UploadFile, upload_token: Annotat
 
     # set appsettings.json
     SETTINGS_FILE = f"{TARGET_DIR}/{next_build_id}/build/appsettings.json"
-    os.remove(SETTINGS_FILE)
+    if os.path.exists(SETTINGS_FILE):
+        os.remove(SETTINGS_FILE)
     os.symlink(APPSETTINGS_FILE, SETTINGS_FILE)
 
     # restart systemd service
